add: csv writer to stdout

+++ syslog-common/src/lib.rs
@@ -4,8 +4,8 @@
 #[derive(Clone, Copy)]
 pub struct SysCallLog {
     pub ts            : u64,
-    pub syscall       : u32,
-    pub pid           : u32,
+    pub syscall       : u64,
+    pub pid           : u64,
     pub pname_bytes   : [u8; 16]
 }
